review
- but `but-ctx` legacy support into `legacy.rs`
