+++ crates/but-ctx/src/legacy.rs
@@ -1,4 +1,5 @@
 use but_core::{RepositoryExt, sync::RepoExclusive};
+use but_error::Code;
 use but_settings::AppSettings;
 use tracing::instrument;
 
@@ -128,6 +129,26 @@ impl Context {
     )> {
         let ws = self.workspace_from_head()?;
         Ok((self.meta()?, ws))
+    }
+
+    /// Return the configured GitButler default target from persisted project metadata.
+    ///
+    /// This is deliberately not derived from the current-`HEAD` workspace projection. When
+    /// `HEAD` is outside the GitButler workspace, projection may produce an ad-hoc workspace,
+    /// infer its target from the checked-out branch's upstream, or clear target metadata when
+    /// the checked-out branch is outside the managed workspace bounds. Legacy compatibility
+    /// flows that re-enter the workspace or operate from outside-workspace states need the
+    /// configured GitButler target instead.
+    pub fn persisted_default_target(
+        &self,
+    ) -> anyhow::Result<but_meta::virtual_branches_legacy_types::Target> {
+        self.meta_inner()?
+            .data()
+            .default_target
+            .clone()
+            .ok_or_else(|| {
+                anyhow::anyhow!("there is no default target").context(Code::DefaultTargetNotFound)
+            })
     }
 
     /// Return a wrapper for metadata that only supports read-only access when presented with the project wide permission
