+++ crates/but-graph/Cargo.toml
@@ -10,6 +10,11 @@ rust-version.workspace = true
 [lib]
 doctest = false
 
+[features]
+## Gate functions we only need to help `gitbutler-` crates along,
+## slated for culling or review before promotion to non-legacy
+legacy = []
+
 [dependencies]
 but-core.workspace = true
 
@@ -22,6 +27,7 @@ tracing.workspace = true
 itertools.workspace = true
 
 [dev-dependencies]
+but-graph = { path = "." , features = ["legacy"] }
 but-meta = { workspace = true, features = ["legacy"] }
 but-testsupport.workspace = true
