+++ crates/but-graph/tests/graph/workspace/legacy.rs
@@ -17,8 +17,8 @@ fn distinguishes_target_base_from_ref_tip() -> anyhow::Result<()> {
         .validated()?
         .into_workspace()?;
 
-    assert_eq!(ws.target_base_oid(), Some(base_id));
-    assert_eq!(ws.target_ref_tip_oid(), Some(target_tip_id));
+    assert_eq!(ws.target_base_commit_id(), Some(base_id));
+    assert_eq!(ws.target_ref_tip_commit_id(), Some(target_tip_id));
     assert_eq!(
         ws.target_ref_name().map(ToString::to_string),
         Some("refs/remotes/origin/main".to_string())
@@ -37,8 +37,8 @@ fn target_helpers_return_none_without_target() -> anyhow::Result<()> {
         .validated()?
         .into_workspace()?;
 
-    assert_eq!(ws.target_base_oid(), None);
-    assert_eq!(ws.target_ref_tip_oid(), None);
+    assert_eq!(ws.target_base_commit_id(), None);
+    assert_eq!(ws.target_ref_tip_commit_id(), None);
     assert_eq!(ws.target_ref_name(), None);
 
     Ok(())
