+++ crates/but-graph/tests/graph/workspace/mod.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "legacy")]
+mod legacy;
 mod merge_base_with_target_branch;
 mod remote_name;
-mod target_helpers;
