+++ crates/gitbutler-branch-actions/Cargo.toml
@@ -28,7 +28,7 @@ but-workspace = { workspace = true, features = ["legacy"] }
 but-rebase.workspace = true
 but-core = { workspace = true, features = ["legacy"] }
 but-project-handle.workspace = true
-but-graph.workspace = true
+but-graph = { workspace = true, features = ["legacy"] }
 but-oxidize.workspace = true
 but-forge.workspace = true
 but-ctx = { workspace = true, features = ["legacy"] }
