+++ crates/gitbutler-edit-mode/Cargo.toml
@@ -13,7 +13,7 @@ export-schema = ["dep:schemars", "dep:but-schemars"]
 doctest = false
 
 [dependencies]
-but-graph.workspace = true
+but-graph = { workspace = true, features = ["legacy"] }
 but-rebase.workspace = true
 but-core.workspace = true
 but-oxidize.workspace = true
