+++ crates/gitbutler-workspace/Cargo.toml
@@ -12,7 +12,7 @@ doctest = false
 [dependencies]
 but-oxidize.workspace = true
 but-core.workspace = true
-but-ctx.workspace = true
+but-ctx = { workspace = true, features = ["legacy"] }
 but-error.workspace = true
 but-meta = { workspace = true, features = ["legacy"] }
